adding popups for key creation

+++ build/metavinci.py
@@ -1,6 +1,6 @@
-from PyQt5.QtWidgets import QApplication, QMainWindow, QLabel, QGridLayout, QWidget, QCheckBox, QSystemTrayIcon, QSpacerItem, QSizePolicy, QMenu, QAction, QStyle, qApp, QVBoxLayout, QPushButton, QDialog
+from PyQt5.QtWidgets import QApplication, QMainWindow, QLabel, QGridLayout, QWidget, QCheckBox, QSystemTrayIcon, QSpacerItem, QSizePolicy, QMenu, QAction, QStyle, qApp, QVBoxLayout, QPushButton, QDialog, QDesktopWidget, QFileDialog, QMessageBox
 from PyQt5.QtCore import QSize
-from PyQt5.QtGui import QIcon
+from PyQt5.QtGui import QIcon, QPixmap
 from pathlib import Path
 import subprocess
 import os
@@ -27,8 +27,6 @@ SERVICE_RUN_FILE = os.path.join(HOME, '.metavinci', 'run.sh')
 SERVICE_START = os.path.join(FILE_PATH, 'service', 'start.sh')
 APP_ICON_FILE = os.path.join(HOME, '.metavinci', 'metavinci.png')
 APP_ICON = os.path.join(FILE_PATH, 'images', 'metavinci.png')
-HVYM_BIN = Path.home() / '.local'/ 'share'/ 'heavymeta-cli'/ 'hvym'
-PRESS_BIN = Path.home() / '.local' / 'share' / 'heavymeta-press' / 'hvym_press'
 DB_PATH = os.path.join(FILE_PATH, 'data', 'db.json')
 FG_TXT_COLOR = '#98314a'
 
@@ -92,12 +90,19 @@ class Metavinci(QMainWindow):
         # Be sure to call the super class method
         QMainWindow.__init__(self)
         self.HOME = os.path.expanduser('~')
-        self.HVYM = os.path.join(self.HOME, '.local', 'share', 'heavymeta-cli', 'hvym')
+        self.PATH = self.HVYM = Path.home() / '.metavinci'
+        self.KEYSTORE = self.PATH / 'keystore.enc'
+        self.ENC_KEY = self.PATH / 'encryption_key.key'
+        self.HVYM = Path.home() / '.local'/ 'share'/ 'heavymeta-cli'/ 'hvym'
+        self.PRESS = Path.home() / '.local' / 'share' / 'heavymeta-press' / 'hvym_press'
         self.FILE_PATH = Path(__file__).parent
+        self.HVYM_IMG = os.path.join(self.FILE_PATH, 'images', 'metavinci.png')
         self.LOGO_IMG = os.path.join(self.FILE_PATH, 'images', 'hvym_logo_64.png')
         self.UPDATE_IMG = os.path.join(self.FILE_PATH, 'images', 'update.png')
         self.INSTALL_IMG = os.path.join(self.FILE_PATH, 'images', 'install.png')
         self.ICP_LOGO_IMG = os.path.join(self.FILE_PATH, 'images', 'icp_logo.png')
+        self.STYLE_SHEET = os.path.join(self.FILE_PATH, 'data', 'style.qss')
+        self.win_icon = QIcon(self.HVYM_IMG)
         self.icon = QIcon(self.LOGO_IMG)
         self.update_icon = QIcon(self.UPDATE_IMG)
         self.install_icon = QIcon(self.INSTALL_IMG)
@@ -108,19 +113,23 @@ class Metavinci(QMainWindow):
         self.private_key = None
         self.refresh_interval = 8 * 60 * 60  # 8 hours in seconds
 
-        self.setMinimumSize(QSize(480, 80))             # Set sizes
+        
+        self.setMinimumSize(QSize(64, 64))             # Set sizes
         self.setWindowTitle("Metavinci")  # Set a title
-        central_widget = QWidget(self)                  # Create a central widget
-        self.setCentralWidget(central_widget)           # Set the central widget
+        self.central_widget = QWidget(self)                 # Create a central widget
+        self.setCentralWidget(self.central_widget)           # Set the central widget
+        self.setWindowIcon(self.win_icon)          # Set the icon
      
         grid_layout = QGridLayout(self)         # Create a QGridLayout
-        central_widget.setLayout(grid_layout)   # Set the layout into the central widget
-        grid_layout.addWidget(QLabel("Application, which can minimize to Tray", self), 0, 0)
+        self.central_widget.setLayout(grid_layout)   # Set the layout into the central widget
+        label = QLabel("", self)
+        label.setPixmap(QPixmap(self.LOGO_IMG))
+        grid_layout.addWidget(label, 0, 0)
      
         # Add a checkbox, which will depend on the behavior of the program when the window is closed
-        self.check_box = QCheckBox('Minimize to Tray')
-        grid_layout.addWidget(self.check_box, 1, 0)
-        grid_layout.addItem(QSpacerItem(0, 0, QSizePolicy.Expanding, QSizePolicy.Expanding), 2, 0)
+        #self.check_box = QCheckBox('Minimize to Tray')
+        #grid_layout.addWidget(self.check_box, 1, 0)
+        #grid_layout.addItem(QSpacerItem(0, 0, QSizePolicy.Expanding, QSizePolicy.Expanding), 2, 0)
     
         # Init QSystemTrayIcon
         self.tray_icon = QSystemTrayIcon(self)
@@ -134,53 +143,119 @@ class Metavinci(QMainWindow):
         '''
         #show_action = QAction("Show", self)
         icp_new_account_action = QAction(self.ic_icon, "New Account", self)
-        icp_change_account_action = QAction(self.ic_icon, "Change Account", self)
-        icp_balance_action = QAction("ICP Balance", self)
-        oro_balance_action = QAction("ORO Balance", self)
-        ckETH_balance_action = QAction("ckETH Balance", self)
-        ckBTC_balance_action = QAction("ckBTC Balance", self)
-        install_hvym_action = QAction(self.install_icon, "Install hvym", self)
-        install_press_action = QAction(self.install_icon, "Install press", self)
-        update_tools_action = QAction(self.update_icon, "Update Tools", self)
-        quit_action = QAction("Exit", self)
         icp_new_account_action.triggered.connect(self.new_ic_account)
+
+        icp_change_account_action = QAction(self.ic_icon, "Change Account", self)
         icp_change_account_action.triggered.connect(self.change_ic_account)
+
+        icp_balance_action = QAction("ICP Balance", self)
         icp_balance_action.triggered.connect(self.get_icp_balance)
+
+        oro_balance_action = QAction("ORO Balance", self)
         oro_balance_action.triggered.connect(self.get_oro_balance)
+
+        ckETH_balance_action = QAction("ckETH Balance", self)
         ckETH_balance_action.triggered.connect(self.get_ckETH_balance)
+
+        ckBTC_balance_action = QAction("ckBTC Balance", self)
         ckBTC_balance_action.triggered.connect(self.get_ckBTC_balance)
-        update_tools_action.triggered.connect(self.update_tools)
+
+        install_hvym_action = QAction(self.install_icon, "Install hvym", self)
         install_hvym_action.triggered.connect(self._install_hvym)
+
+        install_press_action = QAction(self.install_icon, "Install press", self)
         install_press_action.triggered.connect(self._install_press)
 
+        update_tools_action = QAction(self.update_icon, "Update Tools", self)
+        update_tools_action.triggered.connect(self.update_tools)
+
+        quit_action = QAction("Exit", self)
+        quit_action.triggered.connect(qApp.quit)
+
+        gen_keys_action = QAction("Generate Key Pair", self)
+        gen_keys_action.triggered.connect(self.generate_store_keypair)
+
+        import_keys_action = QAction("Import Key Pair", self)
+        import_keys_action.triggered.connect(self.import_keys)
+        
         # Add a new action for tasks
         task_action = QAction("Show Tasks", self)
         task_action.triggered.connect(self.show_tasks_popup)
 
-        quit_action.triggered.connect(qApp.quit)
         tray_menu = QMenu()
         tray_accounts_menu = tray_menu.addMenu("Accounts")
-        tray_tools_menu = tray_menu.addMenu("Tools")
         tray_ic_accounts_menu = tray_accounts_menu.addMenu("IC")
         tray_ic_accounts_menu.addAction(icp_new_account_action)
         tray_ic_accounts_menu.addAction(icp_change_account_action)
+
+
+        tray_tools_menu = tray_menu.addMenu("Tools")
+        tray_tools_menu.addAction(update_tools_action)
+        if not self.HVYM.is_file():
+            tray_tools_menu.addAction(install_hvym_action)
+        if not self.PRESS.is_file():
+            tray_tools_menu.addAction(install_press_action)
+
         tray_balances_menu = tray_ic_accounts_menu.addMenu("Balances")
         tray_balances_menu.addAction(icp_balance_action)
         tray_balances_menu.addAction(oro_balance_action)
         tray_balances_menu.addAction(ckETH_balance_action)
         tray_balances_menu.addAction(ckBTC_balance_action)
+
+        tray_keys_menu = tray_menu.addMenu("Keys")
+        tray_keys_menu.addAction(gen_keys_action)
+        tray_keys_menu.addAction(import_keys_action)
+
         tray_menu.addAction(task_action)
-        tray_tools_menu.addAction(update_tools_action)
-        if not HVYM_BIN.exists():
-            tray_tools_menu.addAction(install_hvym_action)
-        if not PRESS_BIN.exists():
-            tray_tools_menu.addAction(install_press_action)
         tray_menu.addAction(quit_action)
+
         self.tray_icon.setContextMenu(tray_menu)
         self.tray_icon.show()
         self._installation_check()
+        self.setStyleSheet(Path(str(self.STYLE_SHEET)).read_text())
         # self.generate_store_keypair()
         # self.import_keys()
+
+    def center(self):
+        qr = self.frameGeometry()
+        cp = QDesktopWidget().availableGeometry().center()
+        qr.moveCenter(cp)
+        self.move(qr.topLeft())
+
+    def open_dir_dialog(self, prompt):
+        self.show()
+        dir_name = QFileDialog.getExistingDirectory(self.central_widget, prompt)
+        if dir_name:
+            path = Path(dir_name)
+            return str(path)
+        
+        self.hide()
+        
+    def open_file_dialog(self, prompt, filter="*.key"):
+        self.show
+        filename, ok = QFileDialog.getOpenFileName(
+            self,
+            prompt, 
+            self.HOME, 
+            filter
+        )
+        if filename:
+            path = Path(filename)
+            return str(path)
+
+        self.hide()
+
+        
+    def open_confirm_dialog(self, prompt):
+        self.show()
+        result = False
+        response = QMessageBox.question(self, '?', prompt, QMessageBox.Yes | QMessageBox.No, QMessageBox.No)
+
+        if response == QMessageBox.Yes:
+            result = True
+
+        self.hide()
+        return result
      
     def show_tasks_popup(self):
         #Create a dict for popup buttons
@@ -196,6 +271,15 @@ class Metavinci(QMainWindow):
 
     # Task function for generating and storing the app keypair using biscuit-python
     def generate_store_keypair(self):
+        encryption_key_path = self.open_dir_dialog("Select Directory to Store Encryption Key")
+
+        if encryption_key_path == None:
+            return
+        else:
+            encryption_key_path = os.path.join(encryption_key_path, 'encryption_key.key')
+            print(encryption_key_path)
+            print(os.path.isfile(encryption_key_path))
+        
         try:
             # Generate keypair using biscuit-python
             keypair = KeyPair()
@@ -207,9 +291,8 @@ class Metavinci(QMainWindow):
             
             # print("key generated & stored : ",serialized_keys)
             # locate .metavinci directory in user's home directory
-            meta_dir = os.path.join(HOME, '.metavinci')
-            if not os.path.exists(meta_dir):
-                os.makedirs(meta_dir, mode=0o700)  # Create directory with secure permissions
+            if not self.PATH.exists():
+                self.PATH.mkdir(mode=0o700)  # Create directory with secure permissions
 
             # # Generate a symmetric encryption key
             encryption_key = Fernet.generate_key()
@@ -218,41 +301,50 @@ class Metavinci(QMainWindow):
             serialized_keys_str = json.dumps(serialized_keys)
             encrpted_keys = cipher_suite.encrypt(serialized_keys_str.encode())
 
-            keystore_path = os.path.join(meta_dir, 'keystore.enc')
-            encryption_key_path = os.path.join(meta_dir, 'encryption_key.key')
+            keystore_path = str(self.KEYSTORE)
 
-            # # # Write the private key securely (restrict access to owner)
-            with open(keystore_path, 'wb') as keystore_file:
-                keystore_file.write(encrpted_keys)
-            os.chmod(keystore_path, 0o600)
+            if os.path.isfile(encryption_key_path):
+                overwrite = self.open_confirm_dialog(f"File {encryption_key_path} already exists, do you want to overwrite it?")
+                if overwrite == True:
+                    # # # Write the encryption key
+                    with open(encryption_key_path, 'wb') as encrypt_key_file:
+                        encrypt_key_file.write(encryption_key)
 
-            # # # Write the encryption key
-            with open(encryption_key_path, 'wb') as encrypt_key_file:
-                encrypt_key_file.write(encryption_key)
+            if self.KEYSTORE.is_file():
+                overwrite = self.open_confirm_dialog(f"File {keystore_path} already exists, do you want to overwrite it?")
+                if overwrite == True:
+                    # # # Write the private key securely (restrict access to owner)
+                    with open(keystore_path, 'wb') as keystore_file:
+                        keystore_file.write(encrpted_keys)
+                    os.chmod(keystore_path, 0o600)
 
-            # # Show success message
-            _prompt_popup("keypair generated and stored successfully at : "+keystore_path, True)
         except Exception as e:
             # Show error message
             print(e)
             _prompt_popup("Error generating and storing app keypair.")
 
     def import_keys(self):
+        encryption_key_path = self.open_file_dialog("Load encryption key.")
+
+        print(encryption_key_path)
+
+        if encryption_key_path == None:
+            return
+        
         try:
             # locate .metavinci directory in user's home directory
-            meta_dir = os.path.join(HOME, '.metavinci')
-            if not os.path.exists(meta_dir):
-                os.makedirs(meta_dir, mode=0o700)
+            if not self.PATH.exists():
+                self.PATH.mkdir(mode=0o700)
             
             # Read the encrypted keys
-            keystore_path = os.path.join(meta_dir, 'keystore.enc')
-            encrypt_key_path = os.path.join(meta_dir, 'encryption_key.key')
+            keystore_path = str(self.KEYSTORE)
+            encryption_key_path = str(self.ENC_KEY)
 
-            if not os.path.exists(keystore_path) or not os.path.exists(encrypt_key_path):
+            if not os.path.exists(keystore_path) or not os.path.exists(encryption_key_path):
                 _prompt_popup("No keypair found to import.")
                 return
             
-            with open(encrypt_key_path, 'rb') as encrypt_key_file:
+            with open(encryption_key_path, 'rb') as encrypt_key_file:
                 encryption_key = encrypt_key_file.read()
             cipher_suite = Fernet(encryption_key)
 
@@ -266,8 +358,6 @@ class Metavinci(QMainWindow):
             self.private_key = PrivateKey.from_hex(serialized_keys['private_key'])
             self.public_key = PublicKey.from_hex(serialized_keys['public_key'])
 
-            # Show success message
-            _prompt_popup("keypair imported successfully.")
         except Exception as e:
             # Show error message
             print(e)
@@ -425,6 +515,8 @@ class Metavinci(QMainWindow):
     # Override closeEvent, to intercept the window closing event
     # The window will be closed only if there is no check mark in the check box
     def closeEvent(self, event):
+        print('!!!!!!!!!!!!!')
+        print(event)
         if self.check_box.isChecked():
             event.ignore()
             self.hide()
@@ -434,6 +526,7 @@ class Metavinci(QMainWindow):
                 QSystemTrayIcon.Information,
                 2000
             )
+
     def _subprocess(self, command):
         try:
             output = subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
@@ -442,13 +535,13 @@ class Metavinci(QMainWindow):
             return None
 
     def new_ic_account(self):
-        return(self._subprocess(f'{self.HVYM} icp-new-account'))
+        return(self._subprocess(f'{str(self.HVYM)} icp-new-account'))
 
     def change_ic_account(self):
-        return(self._subprocess(f'{self.HVYM} icp-set-account'))
+        return(self._subprocess(f'{str(self.HVYM)} icp-set-account'))
 
     def hvym_check(self):
-        return(self._subprocess(f'{self.HVYM} check'))
+        return(self._subprocess(f'{str(self.HVYM)} check'))
 
     def update_tools(self):
         answer = _choice_popup('You want to update Heavymeta Tools?')
@@ -461,13 +554,15 @@ class Metavinci(QMainWindow):
             loading= _loading_message('UPDATING')
             loading.Play()
             self._update_blender_addon(version)
-            self._update_cli()
-            self._subprocess(f'{self.HVYM} update-npm-modules')
-            self._update_press()
+            if self.HVYM.is_file():
+                self._update_cli()
+                self._subprocess(f'{str(self.HVYM)} update-npm-modules')
+            if self.PRESS.is_file():
+                self._update_press()
             loading.Close()
 
     def _installation_check(self):
-        if not os.path.isfile(self.HVYM):
+        if not self.HVYM.is_file():
             print('Install the cli')
             self._install_hvym()
         else:
@@ -481,17 +576,15 @@ class Metavinci(QMainWindow):
         check = self.hvym_check()
         if installed != None and check != None and check.strip() == 'ONE-TWO':
             print('hvym is on path')
-            print(self.HVYM)
-            self._subprocess(f'{self.HVYM} up')
+            print(str(self.HVYM))
+            self._subprocess(f'{str(self.HVYM)} up')
             self._subprocess('. ~/.bashrc')
         else:
             print('hvym not installed.')
 
     def _delete_hvym(self):
-        home = Path.home()
-        hvym = home / '.local' / 'share' / 'heavymeta-cli' / 'hvym'
-        
-        hvym.unlink
+        if self.HVYM.is_file():
+            self.HVYM.unlink
 
     def _delete_blender_addon(self, version):
         home = Path.home()
@@ -522,10 +615,8 @@ class Metavinci(QMainWindow):
         _subprocess('curl -L https://github.com/inviti8/hvym_press/raw/main/install.sh | bash')
 
     def _delete_press(self):
-        home = Path.home()
-        press = home / '.local' / 'share' / 'heavymeta-press' / 'hvym_press'
-        
-        press.unlink
+        if self.PRESS.is_file():
+            self.PRESS.unlink
 
     def _update_press(self):
         self._delete_press()
@@ -537,7 +628,9 @@ def up():
     import sys
     app = QApplication(sys.argv)
     mw = Metavinci()
-    #mw.show()
+    mw.show()
+    mw.center()
+    mw.hide()
     sys.exit(app.exec())
     click.echo("Metavinci up")
